requirements.txt
flaviabernardes
Django
easy_thumbnails
django_image_cropping
mailchimp
